consultoria-main/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
APScheduler==3.10.4
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
pandas==2.2.2
openpyxl==3.1.5
WeasyPrint==61.2
